Initial commit.
The bin/prancer stuff is somewhat fleshed out, App::Prancer::Handler is waiting
for a proper resolution mechanism, because the trie idea I initially wanted
to go with doesn't quite work well in the face of wildcards.

However, since the handler trait is called at compile-time, I think it behooves
me to do the method resolution right there and then, so the rest of the app
just has to do a quick hash lookup for whatever URLs it wants to resolve.

It's my intention that plugins are all to be done as roles, ideally hanging
off the 'handler' trait. I.E. they all declare a 'handler' trait that just
chains on from whatever the previous handler did.

So as a side effect, handlers shouldn't ever call .wrap() on a method.